requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==3.1.12
djangorestframework==3.11.2
djangorestframework-simplejwt==4.0.0
djongo==1.3.7
idna==3.10
PyJWT==2.10.0
pymongo==3.11.4
pytz==2024.2
requests==2.32.3
sqlparse==0.2.4
typing_extensions==4.12.2
urllib3==2.2.3
